cleaned up sidebar components and naming to all be PascalCase for components and kebab-case for non-components
