+++ components/layout/SidebarConnect.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import { ArrowUpRight } from "lucide-react";
 import { SidebarGroup, SidebarMenuButton } from "@/components/ui/sidebar";
-import { sidebarData } from "@/content/sidebar";
+import { sidebarData } from "@/content/sidebar-data";
 
 export function SidebarConnect() {
     return (
